쿼리 2 results were returned 에러 해결중

+++ src/main/java/shop/mtcoding/filmtalk/showtime/Showtime.java
@@ -27,7 +27,7 @@
         private Screen screen; //상영관
 
 
-        private Double price;
+        private Integer price;
 
 
 
@@ -43,7 +43,7 @@
         }
 
         @Builder
-        public Double getPrice() {
+        public Integer getPrice() {
             return price;
         }
     }
